Update Material color scheme and libraries versions
This commit updates the color scheme in the Navigation Drawer of the Orna Assistant app to match Material Theme standards, improving accessibility. It also upgrades the versions of multiple libraries in the build.gradle file, enhancing the app performance and compatibility. The unnecessary coroutine explanatory comment and the unused text display in the Box component of the Modal Navigation Drawer have been removed for code clean-up purposes.