Add OrnaTower and OrnaGuide web view activities
The commit introduces two new activities - OrnaTowerActivity and OrnaGuideActivity which provide embedded web views for the respective URLs. It also adds buttons to navigate to these new activities. The 'Version Lloir' item is removed from the settings menu and a new permission for posting notifications is included. Furthermore, the viewBinding feature is enabled in build.gradle.