Refactor color scheme and improve layout structure
This commit introduces a new set of colors, resulting in a theme update and modifies various layout and style files to enhance readability as well as structure. Replacing hardcoded values with references from dimens and strings resources paves way to a more maintainable and organized way of managing the app layout.