Refactor MainActivity and improve handling of Accessibility permissions
This commit refactors MainActivity for better maintainability and performance. The TabLayout's selection listener was replaced with a more efficient TabLayoutMediator. Additionally, the manner in which the app requests for Accessibility service permission from the user has been improved to ensure the app will function correctly. Unregistering for shared preference changes is handled within onPause() to improve life-cycle management.